- make sure all RTP ports are unsigned - prevents totally bogus SDP offers from being made due to conversion kabooms (m=audio 4294935774 RTP/AVP 3) (upstream r8160)

+++ Control/L3TranEntry.cpp
@@ -170,7 +170,7 @@ void TranEntry::setDialog(SIP::SipDialog *dialog) { mDialog = dialog; dialog->se
 void TranEntry::txFrame(GSM::AudioFrame* frame, unsigned numFlushed) { getDialog()->txFrame(frame,numFlushed); }
 GSM::AudioFrame *TranEntry::rxFrame() { return getDialog()->rxFrame(); }	// Crashes if rtp not established.
 
-short TranEntry::getRTPPort() const
+unsigned TranEntry::getRTPPort() const
 {
 	if (SipDialog *dialog = getDialog()) { return dialog->RTPPort(); }
 	return 0;
